[TH2-2143] Moved functions related to hashing CR name to the CustomResourceUtils class
